Make it easier to just use Boot’s dependency management with Gradle
Closes gh-11059
