+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/gradle/managing-dependencies/depend-on-plugin-milestone.gradle
@@ -1,11 +1,9 @@
 buildscript {
 	repositories {
-		maven { url 'https://repo.spring.io/libs-snapshot' }
+		maven { url 'https://repo.spring.io/libs-milestone' }
 	}
 
 	dependencies {
 		classpath 'org.springframework.boot:spring-boot-gradle-plugin:{version}'
 	}
 }
-
-apply plugin: 'org.springframework.boot'
