+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/gradle/managing-dependencies/depend-on-plugin-release.gradle
@@ -1,0 +1,3 @@
+plugins {
+	id 'org.springframework.boot' version '{version}' apply false
+}
