+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/gradle/managing-dependencies/depend-on-plugin-snapshot.gradle
@@ -7,5 +7,3 @@ buildscript {
 		classpath 'org.springframework.boot:spring-boot-gradle-plugin:{version}'
 	}
 }
-
-apply plugin: 'org.springframework.boot'
