+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/java/org/springframework/boot/gradle/plugin/DependencyManagementPluginAction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2017 the original author or authors.
+ * Copyright 2012-2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,14 +16,6 @@
 
 package org.springframework.boot.gradle.plugin;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.JarURLConnection;
-import java.net.URL;
-import java.net.URLConnection;
-import java.util.jar.Attributes;
-import java.util.jar.JarFile;
-
 import io.spring.gradle.dependencymanagement.DependencyManagementPlugin;
 import io.spring.gradle.dependencymanagement.dsl.DependencyManagementExtension;
 import org.gradle.api.Action;
@@ -38,48 +30,16 @@ import org.gradle.api.Project;
  */
 final class DependencyManagementPluginAction implements PluginApplicationAction {
 
-	private static final String SPRING_BOOT_VERSION = determineSpringBootVersion();
-
-	private static final String SPRING_BOOT_BOM = "org.springframework.boot:spring-boot-dependencies:"
-			+ SPRING_BOOT_VERSION;
-
 	@Override
 	public void execute(Project project) {
 		project.getExtensions().findByType(DependencyManagementExtension.class)
-				.imports((importsHandler) -> importsHandler.mavenBom(SPRING_BOOT_BOM));
+				.imports((importsHandler) -> importsHandler
+						.mavenBom(SpringBootPlugin.BOM_COORDINATES));
 	}
 
 	@Override
 	public Class<? extends Plugin<Project>> getPluginClass() {
 		return DependencyManagementPlugin.class;
-	}
-
-	private static String determineSpringBootVersion() {
-		String implementationVersion = DependencyManagementPluginAction.class.getPackage()
-				.getImplementationVersion();
-		if (implementationVersion != null) {
-			return implementationVersion;
-		}
-		URL codeSourceLocation = DependencyManagementPluginAction.class
-				.getProtectionDomain().getCodeSource().getLocation();
-		try {
-			URLConnection connection = codeSourceLocation.openConnection();
-			if (connection instanceof JarURLConnection) {
-				return getImplementationVersion(
-						((JarURLConnection) connection).getJarFile());
-			}
-			try (JarFile jarFile = new JarFile(new File(codeSourceLocation.toURI()))) {
-				return getImplementationVersion(jarFile);
-			}
-		}
-		catch (Exception ex) {
-			return null;
-		}
-	}
-
-	private static String getImplementationVersion(JarFile jarFile) throws IOException {
-		return jarFile.getManifest().getMainAttributes()
-				.getValue(Attributes.Name.IMPLEMENTATION_VERSION);
 	}
 
 }
