+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/test/java/org/springframework/boot/gradle/plugin/OnlyDependencyManagementIntegrationTests.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2012-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.boot.gradle.plugin;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import org.springframework.boot.gradle.junit.GradleCompatibilitySuite;
+import org.springframework.boot.gradle.testkit.GradleBuild;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+/**
+ * Integration tests for configuring a project to only use Spring Boot's dependency
+ * management.
+ *
+ * @author Andy Wilkinson
+ */
+@RunWith(GradleCompatibilitySuite.class)
+public class OnlyDependencyManagementIntegrationTests {
+
+	@Rule
+	public GradleBuild gradleBuild;
+
+	@Test
+	public void dependencyManagementCanBeConfiguredUsingCoordinatesConstant() {
+		assertThat(this.gradleBuild.build("dependencyManagement").getOutput())
+				.contains("org.springframework.boot:spring-boot-starter ");
+	}
+
+}
